feat : rename to favorite toon
